+++ src/main/kotlin/io/github/sd3v/mflashcardsbe/repository/entity/FlashcardEntity.kt
@@ -1,3 +1,3 @@
 package io.github.sd3v.mflashcardsbe.repository.entity
 
-data class FlashcardEntity(val id: Long, val front: String, val back: String, val deckId: Int)
+data class FlashcardEntity(val id: Long, val front: String, val back: String, val deckId: Long)
